+++ toddy/src/main.rs
@@ -1,3 +1,0 @@
-fn main() -> iced::Result {
-    toddy::run(toddy_core::app::ToddyAppBuilder::new())
-}